+++ pkgs/test_core/lib/src/runner/configuration.dart
@@ -30,6 +30,20 @@ import 'suite.dart';
 /// The key used to look up [Configuration.current] in a zone.
 final _currentKey = Object();
 
+/// Encapsulates a file/directory path and filter options specific to this path.
+class PathConfiguration {
+  const PathConfiguration({
+    required this.testPath,
+    this.testPatterns,
+  });
+
+  /// The explicit path to a test suite.
+  final String testPath;
+
+  /// Name filters specific to [testPath].
+  final List<Pattern>? testPatterns;
+}
+
 /// A class that encapsulates the command-line configuration of the test runner.
 class Configuration {
   /// An empty configuration with only default values.
@@ -129,8 +143,9 @@ class Configuration {
   final Set<String>? _foldTraceOnly;
 
   /// The paths from which to load tests.
-  List<String> get paths => _paths ?? ['test'];
-  final List<String>? _paths;
+  List<PathConfiguration> get paths =>
+      _paths ?? const [PathConfiguration(testPath: 'test')];
+  final List<PathConfiguration>? _paths;
 
   /// Whether the load paths were passed explicitly or the default was used.
   bool get explicitPaths => _paths != null;
@@ -251,7 +266,7 @@ class Configuration {
       required int? concurrency,
       required int? shardIndex,
       required int? totalShards,
-      required Iterable<String>? paths,
+      required Iterable<PathConfiguration>? paths,
       required Iterable<String>? foldTraceExcept,
       required Iterable<String>? foldTraceOnly,
       required Glob? filename,
@@ -360,7 +375,7 @@ class Configuration {
           int? concurrency,
           int? shardIndex,
           int? totalShards,
-          Iterable<String>? paths,
+          Iterable<PathConfiguration>? paths,
           Iterable<String>? foldTraceExcept,
           Iterable<String>? foldTraceOnly,
           Glob? filename,
@@ -645,7 +660,7 @@ class Configuration {
   factory Configuration.localRunner(
           {required int? pubServePort,
           required Iterable<Pattern>? patterns,
-          required Iterable<String>? paths,
+          required Iterable<PathConfiguration>? paths,
           required Glob? filename,
           required BooleanSelector? includeTags,
           required BooleanSelector? excludeTags,
@@ -734,7 +749,7 @@ class Configuration {
       required int? concurrency,
       required this.shardIndex,
       required this.totalShards,
-      required Iterable<String>? paths,
+      required Iterable<PathConfiguration>? paths,
       required Iterable<String>? foldTraceExcept,
       required Iterable<String>? foldTraceOnly,
       required Glob? filename,
@@ -965,7 +980,7 @@ class Configuration {
       int? concurrency,
       int? shardIndex,
       int? totalShards,
-      Iterable<String>? paths,
+      Iterable<PathConfiguration>? paths,
       Iterable<String>? exceptPackages,
       Iterable<String>? onlyPackages,
       Glob? filename,