+++ pkgs/test_core/lib/src/runner/configuration/load.dart
@@ -352,7 +352,9 @@ class _ConfigurationLoader {
       _validate(pathNode, 'Paths must be relative.',
           (value) => p.url.isRelative(value as String));
 
-      return _parseNode(pathNode, 'path', p.fromUri);
+      return PathConfiguration(
+        testPath: _parseNode(pathNode, 'path', p.fromUri),
+      );
     });
 
     var filename = _parseValue('filename', (value) => Glob(value));