+++ pkgs/test/CHANGELOG.md
@@ -1,5 +1,9 @@
-## 1.18.3-dev
+## 1.19.0-dev
 
+* Support query parameters `name` and `full-name` on test paths, which will
+  apply the filters to only those test suites.
+  * All specified filters must match for a test to run.
+  * Global filters (ie: `--name`) are also still respected and must match.
 * Give a better exception when using `markTestSkipped` outside of a test.
 
 ## 1.18.2